+++ packages/www/src/documentation/components/navigation.ts
@@ -60,6 +60,7 @@ const components: NavigationSection = {
         { title: 'IconButton', path: 'icon-button' },
         { title: 'Link', path: 'link' },
         { title: 'PageSize', path: 'page-size' },
+        { title: 'Pagination', path: 'pagination' },
       ],
     },
     {