+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ActionListItemAction` accepts color prop when value is "danger"
 - `Prompt` component
 - `ActionListManager` component
+- `Pagination` component
 - `ComponentsProvider` component to provide a pre-packaged `ThemeProvider` + `theme`
 
 ### Changed